Fix #10311: make resolve and browser-resolve proper modules, and add tsconfig.json and correct type references
