+++ resolve/resolve-tests.ts
@@ -1,4 +1,4 @@
-/// <reference path="resolve.d.ts" />
+/// <reference types="resolve" />
 
 import * as fs from 'fs';
 import * as resolve from 'resolve';
